move waitForTasks to reindexHelper
