+++ test/unit/int_algolia/scripts/algolia/helper/reindexHelper.test.js
@@ -25,11 +25,13 @@ const mockMoveIndex = jest.fn().mockReturnValue({
         }
     }
 });
+const mockWaitTask = jest.fn();
 jest.mock('*/cartridge/scripts/algoliaIndexingAPI', () => {
     return {
         deleteIndex: mockDeleteIndex,
         copyIndexSettings: mockCopySettingsFromProdIndices,
         moveIndex: mockMoveIndex,
+        waitTask: mockWaitTask,
     }
 }, {virtual: true});
 
@@ -59,4 +61,17 @@ test('moveTemporariesIndices', () => {
     reindexHelper.moveTemporariesIndices('products', ['fr', 'en']);
     expect(mockMoveIndex).nthCalledWith(1, 'test_index___products__fr.tmp', 'test_index___products__fr');
     expect(mockMoveIndex).nthCalledWith(2, 'test_index___products__en.tmp', 'test_index___products__en');
+});
+
+test('waitForTasks', () => {
+    mockWaitTask
+        .mockReturnValue({
+            ok: true,
+            object: { body: { status: 'published' }}
+        });
+    reindexHelper.waitForTasks({ 'testIndex': 33, 'testIndex2': 51 });
+
+    expect(mockWaitTask).toHaveBeenCalledTimes(2);
+    expect(mockWaitTask).toHaveBeenCalledWith('testIndex', 33);
+    expect(mockWaitTask).toHaveBeenCalledWith('testIndex2', 51);
 });
